requirements.txt
pymediainfo
gpustat
requests
plexapi
loguru
rich
fastapi
uvicorn
pynvml
amdsmi
python-dotenv
jinja2
